replace devldap with glauth
